# fastapi-rest-api-guide/openapi.yaml
components:
  schemas:
    DocumentInput:
      properties:
        format:
          example: pdf
          title: Format
          type: string
        template:
          anyOf:
          - type: string
          - type: 'null'
          example: business-report
          title: Template
        title:
          example: Quarterly Report
          title: Title
          type: string
      required:
      - title
      - format
      title: DocumentInput
      type: object
    DocumentOutput:
      properties:
        format:
          title: Format
          type: string
        id:
          title: Id
          type: string
        status:
          default: completed
          title: Status
          type: string
        template:
          anyOf:
          - type: string
          - type: 'null'
          title: Template
        title:
          title: Title
          type: string
      required:
      - id
      - title
      - format
      - template
      title: DocumentOutput
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: A demo API for a technical writing portfolio.
  title: Portfolio Document API
  version: 1.0.0
openapi: 3.1.0
paths:
  /documents:
    get:
      description: List all available documents.
      operationId: list_documents_documents_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DocumentOutput'
                title: Response List Documents Documents Get
                type: array
          description: Successful Response
      summary: List Documents
    post:
      description: Create a new document.
      operationId: create_document_documents_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentInput'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentOutput'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Document
  /documents/{document_id}:
    delete:
      description: Delete a document by its ID.
      operationId: delete_document_documents__document_id__delete
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Document
    get:
      description: Retrieve a single document by its ID.
      operationId: get_document_documents__document_id__get
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentOutput'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Document
  /health:
    get:
      description: Health check endpoint.
      operationId: health_check_health_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Health Check